INTRODUCE:
In Daily, there are a lot of information need to be handled. Feeling that there are still having trouble and something hardly to control in the management in kindergarten so we try to create a management agency to make is become more easily and simply. We use HTML, CSS to design graphic, SQL to manage and update data, PHP to combile SQL, HTML, CSS to show all system's functions